requirements.txt
click==8.1.7
lxml==5.1.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
python-docx==1.1.0
pytoolconfig==1.3.1
rope==1.12.0
typing_extensions==4.10.0
